Move user create page component (#178)
